Generate `.d.ts` files using `rolldown-plugin-dts`.
