+++ packages/load-source-map/package.json
@@ -5,14 +5,13 @@
   "type": "module",
   "license": "MIT",
   "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.cjs"
-    },
+    ".":  "./dist/index.js",
     "./package.json": "./package.json"
   },
   "scripts": {
-    "build": "rolldown -c && tsc"
+    "build": "rolldown -c"
+  },
+  "devDependencies": {
+    "rolldown-plugin-dts": "^0.10.1"
   }
 }
