+++ packages/sonda/package.json
@@ -61,8 +61,8 @@
     "prepack": "pnpm run test && pnpm run build && clean-package -rm devDependencies scripts",
     "build:load-source-map": "cd ../load-source-map && pnpm run build",
     "build:html": "cd ../html-report && pnpm run build",
-    "build:dist": "rolldown -c && tsc",
-    "build": "pnpm run build:load-source-map && pnpm run build:dist && pnpm run build:html",
+    "build:dist": "rolldown -c",
+    "build": "pnpm run build:load-source-map && pnpm run build:dist",
     "dev": "rolldown -w -c"
   },
   "dependencies": {
@@ -77,6 +77,7 @@
     "esbuild": "^0.25.3",
     "load-source-map": "workspace:^",
     "next": "^15.3.1",
+    "rolldown-plugin-dts": "^0.10.1",
     "rollup": "^4.40.1",
     "vite": "^6.3.4",
     "vitest": "^3.1.2",
