+++ packages/sonda/src/entrypoints/nuxt.ts
@@ -1,5 +1,4 @@
-import VitePlugin from './vite.js';
-import { Config, type UserOptions } from '../index.js';
+import { SondaVitePlugin, Config, type UserOptions } from 'sonda';
 import type { NuxtModule, Nuxt } from '@nuxt/schema';
 
 export default function SondaNuxtPlugin( userOptions: UserOptions = {} ): NuxtModule {
@@ -29,7 +28,7 @@ export default function SondaNuxtPlugin( userOptions: UserOptions = {} ): NuxtMo
 
       // Add the Sonda plugin to the Vite configuration
       config.plugins ??= [];
-      config.plugins.push( VitePlugin( sondaOptions ) as any );
+      config.plugins.push( SondaVitePlugin( sondaOptions ) as any );
     } )
   }
 }
