+++ packages/sonda/src/entrypoints/vite.ts
@@ -1,19 +1,3 @@
-import RollupPlugin from './rollup.js';
-import { Config, type UserOptions } from '../index.js';
-import type { PluginOption } from 'vite';
+import { SondaVitePlugin } from 'sonda';
 
-export default function VitePlugin( userOptions: UserOptions = {} ): PluginOption {
-  const options = new Config( userOptions, {
-    integration: 'vite'
-  } );
-
-  if ( !options.enabled ) {
-    return { name: 'sonda-vite' };
-  }
-
-  return {
-    ...RollupPlugin( options ),
-    name: 'sonda-vite',
-    apply: 'build'
-  };
-}
+export default SondaVitePlugin;
