+++ packages/sonda/src/report/formatters/JsonFormatter.ts
@@ -1,5 +1,5 @@
 import { Formatter } from './Formatter.js';
-import type { JsonReport } from '../producer.js';
+import type { JsonReport } from '../types.js';
 
 export class JsonFormatter extends Formatter {
 	protected extension = '.json';
