+++ packages/sonda/tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"extends": "../../tsconfig.json",
 	"compilerOptions": {
-		"rootDir": "./src",
+		"rootDir": "./src"
 	},
 	"include": [
 		"src"
